chore: docs and TODOs
