+++ .gitignore
@@ -6,3 +6,6 @@ zingolib/proptest-regressions
 test_binaries/bins/*
 libtonode-tests/tests/chain_generics.proptest-regressions
 libtonode-tests/store_all_checkpoints_test
+
+.DS_Store
+*.dat
